Add basic tests for the session
