+++ transport/session/Cargo.toml
@@ -20,8 +20,8 @@ futures = { workspace = true }
 libp2p-identity = { workspace = true }
 thiserror = { workspace = true }
 
-core-path = { workspace = true }
 hopr-internal-types = { workspace = true }
 
 [dev-dependencies]
 async-std = { workspace = true }
+mockall = { workspace = true }
